flask/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.1
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
openai==1.26.0
pydantic==2.7.1
pydantic_core==2.18.2
PyPDF2==3.0.1
python-docx==1.1.2
regex==2024.4.28
scikit-learn==1.4.2
scipy==1.13.0
sniffio==1.3.1
SQLAlchemy==2.0.30
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.11.0
Werkzeug==3.0.3
